fix and update user tests

+++ packages/medusa/src/services/user.ts
@@ -3,6 +3,7 @@ import { MedusaError } from "medusa-core-utils"
 import Scrypt from "scrypt-kdf"
 import { EntityManager } from "typeorm"
 import { TransactionBaseService } from "../interfaces"
+import AnalyticsFeatureFlag from "../loaders/feature-flags/analytics"
 import { User } from "../models"
 import { UserRepository } from "../repositories/user"
 import { FindConfig } from "../types/common"
@@ -12,15 +13,17 @@ import {
   UpdateUserInput,
 } from "../types/user"
 import { buildQuery, setMetadata } from "../utils"
+import { FlagRouter } from "../utils/flag-router"
 import { validateEmail } from "../utils/is-email"
 import AnalyticsConfigService from "./analytics-config"
 import EventBusService from "./event-bus"
 
 type UserServiceProps = {
   userRepository: typeof UserRepository
-  analyticsService: AnalyticsConfigService
+  analyticsConfigService: AnalyticsConfigService
   eventBusService: EventBusService
   manager: EntityManager
+  featureFlagRouter: FlagRouter
 }
 
 /**
@@ -36,14 +39,24 @@ class UserService extends TransactionBaseService {
 
   protected manager_: EntityManager
   protected transactionManager_: EntityManager
-  protected readonly analyticsService_: AnalyticsConfigService
+  protected readonly analyticsConfigService_: AnalyticsConfigService
   protected readonly userRepository_: typeof UserRepository
   protected readonly eventBus_: EventBusService
+  protected readonly featureFlagRouter_: FlagRouter
 
-  constructor({ userRepository, eventBusService, manager }: UserServiceProps) {
-    super({ userRepository, eventBusService, manager })
+  constructor({
+    userRepository,
+    eventBusService,
+    analyticsConfigService,
+    featureFlagRouter,
+    manager,
+  }: UserServiceProps) {
+    // eslint-disable-next-line prefer-rest-params
+    super(arguments[0])
 
     this.userRepository_ = userRepository
+    this.analyticsConfigService_ = analyticsConfigService
+    this.featureFlagRouter_ = featureFlagRouter
     this.eventBus_ = eventBusService
     this.manager_ = manager
   }
@@ -239,7 +252,8 @@ class UserService extends TransactionBaseService {
   async delete(userId: string): Promise<void> {
     return await this.atomicPhase_(async (manager: EntityManager) => {
       const userRepo = manager.getCustomRepository(this.userRepository_)
-      const analyticsServiceTx = this.analyticsService_.withTransaction()
+      const analyticsServiceTx =
+        this.analyticsConfigService_.withTransaction(manager)
 
       // Should not fail, if user does not exist, since delete is idempotent
       const user = await userRepo.findOne({ where: { id: userId } })
@@ -248,7 +262,13 @@ class UserService extends TransactionBaseService {
         return Promise.resolve()
       }
 
-      await analyticsServiceTx.delete(userId)
+      if (this.featureFlagRouter_.isFeatureEnabled(AnalyticsFeatureFlag.key)) {
+        try {
+          await analyticsServiceTx.delete(userId)
+        } catch (error) {
+          console.warn(error)
+        }
+      }
 
       await userRepo.softRemove(user)
